Refactor module code
